third_party/WebKit/LayoutTests/http/tests/media/autoplay/document-user-activation-feature-policy-alternating.html

<!DOCTYPE html>
<html>
<title>Test that autoplay works on child frames, but only if allowed.</title>
<script src="/js-test-resources/js-test.js"></script>
<script src="/media-resources/media-file.js"></script>
<script src="resources/video-test.js"></script>
<iframe allow="autoplay" src="http://localhost:8000/media/autoplay/resources/nested-iframe-1b.html"></iframe>
<script>
  simulateMouseDown();
  runTest();
</script>
</html>
